avoid implicit type case (#4098)

+++ nav2_smac_planner/include/nav2_smac_planner/node_hybrid.hpp
@@ -406,7 +406,7 @@ public:
   static float getObstacleHeuristic(
     const Coordinates & node_coords,
     const Coordinates & goal_coords,
-    const double & cost_penalty);
+    const float & cost_penalty);
 
   /**
    * @brief Compute the Distance heuristic
